requirements.txt
Django==2.2.4
django_rest_swagger==0.3.10
djangorestframework==3.12.4
